Bossと戦う前に演出でるように

+++ プロジェクトファイル/Engine/ResourceManager/Image.h
@@ -86,7 +86,7 @@ namespace Image
 	//アルファ値を指定
 	//引数：handle	設定したい画像の番号
 	//引数：alpha アルファ値（不透明度）　0～255で指定
-	void SetAlpha(int handle, int alpha);
+	void SetAlpha(int handle, float alpha);
 
 	//アルファ値をゲットする
 	float GetAlpha(int handle);
